Writing a better octicon cache test
