+++ test/components/octicon_component_test.rb
@@ -102,20 +102,4 @@ class PrimerOcticonComponentTest < Minitest::Test
 
     assert_selector("svg.octicon-star path[d]")
   end
-
-  def test_cache_evacuates_after_limit_reached
-    Primer::Octicon::Cache.clear!
-    Primer::Octicon::Cache.stub :limit, 3 do
-      # Assert the limit is stubbed properly
-      assert_equal Primer::Octicon::Cache.limit, 3
-      # Assert the cache is empty
-      assert_equal 0, Primer::Octicon::Cache::LOOKUP.size
-
-      # Preload the cache should be 20 items
-      Primer::Octicon::Cache.preload!
-
-      # Assert the cache size is 3 because the limit is 3
-      assert_equal 3, Primer::Octicon::Cache::LOOKUP.size
-    end
-  end
 end
